// src/lib/enum.ts
export enum SchoolStatus {
  CONTINUING = "CONTINUING",
  COMPLETED = "COMPLETED",
  DROPPED = "DROPPED",
}

export enum LanguageProficiency {
  A1 = "A1", // Beginner
  A2 = "A2", // Elementary
  B1 = "B1", // Intermediate
  B2 = "B2", // Upper Intermediate
  C1 = "C1", // Advanced
  C2 = "C2", // Proficient/Near-native
  NATIVE = "NATIVE", // Native speaker
}

export enum CVSection {
  ABOUT,
  SKILLS,
  EXPERIENCES,
  PROJECTS,
  EDUCATION,
  CERTIFICATES,
  LANGUAGES,
}